requirements.txt
imgurpython==1.1.7
progressbar2==3.12.0
python-utils==2.0.1
requests==2.11.1
six==1.10.0
